// src/api/services/base/models/createRiskStudentsDetailDto.ts
/**
 * Generated by orval v6.12.1 🍺
 * Do not edit manually.
 * EAS
 * The EAS API description
 * OpenAPI spec version: 0.1
 */

export interface CreateRiskStudentsDetailDto {
  facultyName: string;
  facultyEmail: string;
  courseNumber: string;
  courseCrn: string;
  studentName: string;
  studentId?: string;
  studentNetId?: string;
  notes?: string;
  isGraduateStudent: boolean;
  graduateProgram?: string;
  sendEmail?: boolean;
  factors: string[];
}